back-main/requirements.txt
flask==3.0.2
streamlit==1.32.0
python-dotenv==1.0.1
flask-cors==4.0.0
pandas==2.2.1
pillow==10.2.0
plotly==5.18.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27